+++ src/main/java/zd/zero/waifu/motivator/plugin/personality/EmotionDerivationUnit.kt
@@ -1,0 +1,10 @@
+package zd.zero.waifu.motivator.plugin.personality
+
+import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvent
+
+internal interface EmotionDerivationUnit {
+    fun deriveEmotion(
+        motivationEvent: MotivationEvent,
+        emotionalState: EmotionalState
+    ): EmotionalState
+}
